spoon_toolkits/desearch/requirements.txt
desearch-py>=1.0.0
fastmcp>=0.1.0
python-dotenv>=1.0.0 